fix(AutoSql): fix update sql cannot convert keys
Changed implemention of converting. Not relating to
UNDERSCORE-TO-CAMEL-CASE configuration any more. If passing a JavaBean
as parameter. Will try to match every column to the most likely fields.
(@Column has highest level, then common transforms which where supported
in google-guava CaseFormat)

Another update is stopped converting column
names to lowercase when the parameter is Map type